fix data type too short for node ids
git-svn-id: http://svn.code.sf.net/p/keepright/code@696 4e67cf90-0d29-4caa-83dc-599f2b650fb7

+++ checks/0310_roundabouts.php
@@ -17,7 +17,7 @@ not enough roads connected (normally 3)
 query("DROP TABLE IF EXISTS _tmp_roundabout_parts", $db1);
 query("
 	CREATE TABLE _tmp_roundabout_parts (
-		part int,
+		part bigint,
 		sequence_id int,
 		way_id bigint NOT NULL,
 		first_node_id bigint,
